Widgets: tweak inactive colors
for both Button, LCDSPinBox,  and LCDCombo the styleSheet is not updated anymore when disabling the object but the disabled version is directly included into the sheet.

Instead of using window color and window text color the widget do now pick a darker/lighter color relative to their current settings. This integrates the inactive widgets much better into the GUI.