PreferencesDialog: refactor + tooltips
the individual driver info settings have been put into dedicated functions. They were used both when directly selected and via the Auto option but the code in the two instances did already start the diverge. In addition, everything will be deactivated in case the Auto driver is starting the NullDriver.

- tooltips were added for sample rate and buffer size when JACK is selected