select currently loaded file in AudioFileBrowser
Another convenience layer was build into the default path selection of the AudioFileBrowser (which is used for loading the PlaybackTrack and the Layer files). If "Edit" (PlaybackTrack) or "load layer" is clicked while there is already a sample present, the dialog will select this file as default. If none is present yet, the browser will ressort to the last opened directory stored in the preferences. And if this one was not set yet, the home folder will be used