fix segfault when loading invalid song
on startup using the -s option, providing an invalid name, and sending a MIDI or OSC command immediately.

What happens is the following: Hydrogen can not load the selected song - since it is invalid - and displays a popup dialog which tells that the song could not be loaded. Only after this popup is closed an empty song will be set for Hydrogen::m_pSong. All incoming MIDI and OSC - which the core is already able to handle - will be executed on a pSong == nullptr.

I introduced checks on all API endpoints for empty song.