tweak METRONOME_EVENT handling + Director UI & routing
Triggering/updating of the Director was previously done in a outdated and error-prone way. It did not react on event of the core but solely EventQueue events issued by the GUI (?!). To accomplish this the METRONOME_EVENT did contain not just its two "natural" states of 0 - bar and 1 - emphasised (first) bar but also two more representing an update of the tags (only issued when switching songs!) and an update of the song. Now, the Director does directly listen to `SONG_UPDATE_EVENT` and `TIMELINE_UPDATE_EVENT` and is updated every time a tag is changed.

In addition, the hardcoded colors have been dropped in favor of those defined in the ColorTheme.