GUI: disentangle song update process
Previously, the HydrogenApp did react on both the update song and quit event by doing some stuff and telling a lot of other components of the GUI to update as well. This quite complex information flow did grew naturally and is, unfortunately, not that easy to debug. Instead, now the individual parts of the GUI do directly listen to the UPDATE_SONG_EVENT and QUIT_EVENT and do their update stuff without interacting with one another. (Attention: this probably just disentangled _some_ part of the song load process and not the whole thing altogether)