fix window title for new song
I did touch the default song name when reworking the empty/default song retrieval. This caused the window name to be off when creating a new Song.

While at it I also fixed #687: For a new song, "Untitled Song" or a different Song::m_sName if set will be shown (since there is no file associated with an empty song.). If Song::m_sName was not touched by the user but the song already has a Song::m_sFilename (it was saved to disk) the file name including suffix will be shown (previous behavior). If, however, the complete basename of the song file and Song::m_sName do differ the window title now displays them both as "SONG NAME [FILE_NAME.h2song]". This should point users immediately to any mismatch