Skin: update QLineEdit + add QTextEdit color
background color of QLineEdit was changed to the brighted widget color (whichs purpose is to be distinct from the windowColor the QLineEdit is embedded into). This makes edible regions in various properties dialogs way more easy to spot.

Also, color and background of QTextEdit has been aligned to the one of QLineEdit for a more consistent look