fix song save
When using "save as" to store the current song under a different filename via the GUI, the new filename was not inserted into the list of recently used filenames and not set as lost song filename in the Preferences. In addition, the SONG_SAVE_AS OSC command did only store the current song under a different filename but did not overwrite Song::m_sFilename. Subsequent SONG_SAVE OSC commands would have stored changes in the previous file.