drumkit load: less log messages
Previously, on successful load (at least) two messages where logged for each drumkit: "load drumkit X" and that it is valid with respect to the XSD file. The latter one has been dropped as there would be warning and error logs if the drumkit was not valid. Also, the debug messages displayed if no image is present in a drumkit was dropped. This was neither a common thing to be shipped with the drumkit nor something important.

Loads of DEBUGLOG messages concerning the pan import did flood the log. They were present because in the current implementation the drumkit is allowed to "pan_L" and "pan_R" values or a single "pan" value for each instrument. Thats why drumkits containing the former pair are still valid and not upgraded. However, the debug messages were display as soon as "pan" was not found. They have been dropped.