fix segfault in #1474
That one was quite weird. It had nothing to do with the instrument/drumkit cleanup but was due to an access of an array in the Mixer without any boundchecks. But it did only happen when shutting down the application.

When handing a Note to the Sampler the AudioEngine queues a EVENT_NOTEON and provides the instrument number of the current note. The latter is aquired by comparing Note::__instrument with all the instruments of a song. However, if the Drumkit was recently switched to a smaller one, some instruments at the end of the songs instrument list are removed and all their corresponding notes being purged from the patterns. But there is still an intermediate buffer that is missed in here: the notes of the purged instruments might been already added to the AudioEngine::m_songNoteQueue but not being handed to the sampler (due to the lookahead). The instruments of these notes do not reside in the Song::m_pInstrumentList but in Hydrogen::__instrument_death_row. Therefore, when handing the notes to the Sampler the AudioEngine can not find the corresponding instrument in Song::m_pInstrumentList and triggers an EVENT_NOTEON with a value of `-1`. This will be used in the Mixer as index for a C-style array without further checks and causes a segfault.

I have added both boundary checks for the provided indices in Mixer as well as checking whether the instrument could be found before triggering the EVENT_NOTEON