less conservative instrument removal
if a drumkit was loaded with less instruments than the original one and the user decided to discard the additional ones, those instruments were either deleted right away, on the next drumkit load with less instruments+removal or when quitting Hydrogen. As this just affects notes that are either played by the Sampler or queued in the AudioEngine::m_songNoteQueue the moment the drumkit switch occurs, this behavior is quit conservative. The notes are almost certainly flushed in the next seconds and there is no reason to hold all samples of the instrument till Hydrogen is terminated. Instead, Hydrogen will now check every time the transport is stopped whether there are instruments in the Hydrogen::__instrument_death_row that can be removed (are not queued anymore).