fix drumkit export
In #1490 I messed up the drumkit export via the GUI a little because I misunderstood the way the former implementation was meant to work.

There is now a bigger change in the export UX: Previously, when selecting format version >=0.9.7 the component spinbox was locked to the first component but not hidden which - at least for me - suggest the export of a single component. In the older format the user was able to alter the index of the spinbox. Now, the spinbox is active regardless of the format selected. For the recent format the default option is "All" indicating that all components will be exported. But the user is free to select a single one and export a valid drumkit containing only this component. The former format works as before.

In addition there are two minor changes: when opening the export dialog the cursor is changed to the wait cursor since the list of drumkits gets updated and (depending on how many are installed) this can take a few seconds. Also, when export the drumkit to an existing file the user is prompted whether she wants to overwrite it.