Drumkit export: integrate component and version in filename
Since the user is not allowed to choose the name of the resulting .h2drumkit file herself - just the folder it will be stored to - exporting different versions, like full kit, individual components, and a legacy version, caused each export to overwrite the other. To prevent this from happening and provide a more descriptive file name, the component name is appended to the drumkit name using "_" and in case an export for Hydrogen versions <= 0.9.6 is requested an additional "_legacy" too.

To make things consistent, the exported drumkits will have their drumkit name matching the base name of the compressed kit, e.g. GMRockKit_Main_legacy.