DK export: fix file name validation
The validation of the base name to use for the resulting .h2drumkit was moved to the Filesystem class as other classes might have a use for it too. In addition, it now allows dashes to remain in the string (as I do not see how they are harmful and they are already present in a number of kits), removes dots (as their might be some code left using the base name of the file without checking whether this matches the completeBaseName() in Qt5 terms), and replaces whitespaces by underscores (to improve readability of the resulting filename).

Also, the component name integrated in the filename is also validated as this could be another vector to mess up the export.