DK export: retrieve DK from stored list
SoundLibraryExportDialog maintains a list of all available drumkits (loaded without samples). When exporting a drumkit via the GUI it is not loaded by name but, again, picked from this list.

Loading it directly would have the benefit to be sure to use the most recent version but the name of the drumkit is not something well defined within Hydrogen. Drumkit::load_by_name is expecting the name of the drumkit folder containing both the samples and drumkit.xml files. However, the name property stored in the latter, which is also used as Drumkit::__name can differ (and they actually do for a number of kits we host at SourceForge). Therefore, it is important to retrieve the kit from the list of drumkits used to create the different choices presented in the GUI.