PreferencesDialog: display ALSA device in info
the AlsaAudioDriver tries to establish a connection to the device selected by the user via the Preferences. If this fails, it tries instead to connect to the "default" device which usually works. This change of devices is, however, only shown in the log and users may easily miss it. Now, the currently connected ALSA device is shown in the info section on the right of the audio driver tab.