Fix crash on changing song with Automation Path showing
The automation path keeps a pointer to the song's AutomationPath.
This pointer is updated by the Song Editor, but only as a result
of the event queue, which is asynchronous (and also potentially
lossy). If some event (such as a paint event caused to repair the
display after a confirmation dialog is closed...) is delivered to
the view widget after the song is changed (and the old song destroyed)
but before the Song Editor has processed the update event, the old
pointer will be dereferenced, likely causing a crash.

This change fixes this race condition by explicitly updating the
saved pointer from the song at the start of asynchronously delivered
events. This doesn't feel quite right but is at least consistent
with other editors where state is updated directly from the Song.
One alternative approach would be to use a shared_ptr in the Song
and the editor to ensure that the object at least still exists when
it's referenced in such cases, but that feels further from the right
fix as it presents inconsistent state.