AlsaAudioDriver: fix processing thread
I was lucky enough to get the ALSA driver in such a bricked state that both the transport/playback was not working anymore and Hydrogen did freeze when trying to set a different audio driver in the PreferencesDialog. The latter is indeed a quite serious problem. In case the user sets a "wrong" configuration, it can not be changed via the Preferences anymore but, instead, has to be manually reverted in the hydrogen.conf file. After some debugging I found that the processing thread of the ALSA driver did freeze after two cycles. The state of the playback stream itself was still nice but the corresponding resource "temporary unavailable". Since the device is opened in blocking mode in AlsaAudioDriver::connect() the process thread did freeze at snd_pcm_writei(). The call to AlsaAudioDriver::disconnect() (required to restart/start a new audio driver) is waiting for the thread to finisih which is not going to happen. Thats why Hydrogen did freeze at both shutdown and restarting the driver in the Preferences.

One possible way to fix this is to open the device in non-blocking mode. But this would be quite an invasive change and I am not too familiar with ALSA. So, instead, I introduced a call to the snd_pcm_wait() function which returns as soon as the opened device is ready to process data or returns with an error after a provided timeout. AFAIU this is also done internally in snd_pcm_writei() in blocking mode and will thus only introduce a minimal overhead. On the other hand, it is able to discard the audio output in case the ALSA device is not working anymore. Instead of freezing, it now attempts to annoy the user with error log messages as well as flashing XRun indications and status messages in the GUI in order to indicate a change of the audio driver is required.

I used 100 milliseconds for the timeout. It is enough to make the transport run way to slow to indicate a problem. Smaller values of e.g. 10ms did cause glitches with working ALSA devices.

In addition, I made the ALSA driver use our XRUN event and used the snd_pcm_recover() functions over our custom alsa_xrun_recovery() as the former is part of the ALSA API and recovers from one more error code than our custom one.