Logger: introduce flushing
a Logger::flush() function was introduced which returns once the Logger thread popped all messages in the logger queue (or after 10s). This is used in the fixture of the audio engine unit tests. When enabling log messages in the AudioEngine there are easily so many messages the logger is not finished when the TestRunner already returned. As a result the summary of the test run is one somewhere hidden in the log messages and one needs to scroll to see the actual results