AudioEngine: fix frame <-> tick conversion
within computeTickFromFrame() anmd computeFrameFromTick() the tick size is now calculated in double precision in order to get more accurate results

in computeFrameFromTick() processing of the segments between tempo markers have slightly been changed. If transport is located to the exact position of a tempo marker (provided in ticks) the corresponding frame value might overshoot the position of the tempo marker, is floored and the mismatch is converted to m_fTickMismatch. This latter conversion, however, was previously done using the tick size encountered _after_ the tempo marker (as the frame position was overshooting). When trying to combine the frame position with the tick mismatch in order to retrieve the original tick value things did not end up right since now the m_fTickMismatch is handled as if it is located _before_ the tempo marker. To fix this, the mismatch is handled with the tick size _before_ the tempo marker in the first place.