AudioEngine: handleDriverChange + tweak frame<->tick conversion
Whenever the audio driver is restarted via the PreferencesDialog it may hold a different sample rate. This, however, does affect the tick size computation and invalidates all frame-based values stored in the audio engine, sampler, and processed notes. The AudioEngine::handleDriverChange function will now update all affected values.

There was a slight rounding mismatch in the calculation of the song size in frames in computeTickFromFrame() and computeFrameFromTick(). It was only minor but whenever a relocation was done to a very large value, like 1000 times the song size, the conversion between tick and frames produced a noticable mismatch.