move Sample selection into Note class
previously the Note selection was done inline and with a lot of duplicated code inside the Sampler. This logic has been moved to the Note class in order to make it reusable.

The lead&lag checks in the AudioEngine unit tests have been relaxed and the check for audio glitches does now work for other sample rates too.