fix playing patterns handling
The update of the AudioEngine::m_pPlayingPatterns vector has been refactored into a separate function and is now also performed when relocating transport. This way the transport position in Song mode and the list of patterns being played back is always in sync.

This results in two changes of the UX
- it is now sufficient to relocate to a specific column in song mode in order to add the patterns there to the currently playing ones. Switching back to pattern mode with stacked mode enabled let you play back this selection right away.
- When middle-clicking a pattern in order for it to be played back one can start its playback immediately by pressing the fast forward button or changing to a tick bigger than the current pattern size in general.

In adidtion, the deletion and adding of patterns has been moved into the CoreActionController entirely. This was done to ensure the patterns are cleaned up properly and fixed a couple of bugs
- loading a pattern, start playing it back, and removing it via an Undo action was not cleaned up properly in all instances it occured in the GUI. In one case it resulted in a segfault
- adding the loaded pattern to AudioEngine::m_pNextPattern by middle-clicking it and then removing the pattern using Undo was never handled and always resulted in a segfault.

Also, adding a virtual pattern in stacked pattern mode did not result in adding all patterns contained in the virtual one. This was fixed as well.

Implementational details:
- The COLUMN_CHANGED_EVENT was dropped in favor for PATTERN_CHANGED_EVENT
- Hydrogen::sequencer_setNextPattern() and Hydrogen::sequencer_setOnlyNextPattern() have been renamed and moved to AudioEngine::toggleNextPattern() and AudioEngine::flushAndAddNextPattern()
- AudioEngine::m_pPlayingPatterns and AudioEngine::m_pNextPatterns are now only provided by the AudioEngine as const pointers to other classes. All their handling is done in AudioEngine