AudioEngine: fix PatternStartTick
The initialization of AudioEngine::m_nPatternStartTick with -1 was wrong and is only useful for the corresponding local variable in updateNoteQueue() to indicate it has to be updated. This caused the pattern tick position to be off by one tick in general and a couple of (minor) artifacts. It is now initialized and reset to 0.

AudioEngine::updatePlayingPatterns() is now not called in AudioEngine::updateTransport() anymore but in AudioEngine::locate* . Calling it in the former context could result in the playing patterns to be out of sync since it does not incorporate the lookahead as updateNoteQueue() does. TBH, I am growing suspicious on the whole lookahead concept. Anyway, via the call in locate*() playing patterns are in sync even if the transport is not rolling.