introduce full-height playhead
The playhead is now extended to cover the SongEditor as well as the PatternEditor and the NotePropertiesRuler as well. It comes with its own color, which can be customize in the Preferences.

In addition there are two UX changes:
- clicking the song editor ruler is now alway possible. By doing so transport is relocated to a specific column and transport mode is changed into Song::Mode::Song
- clicking the pattern editor ruler is now also supported. It will relocate transport to a specific tick within the pattern and changes transport mode into Song::Mode::Pattern

All common rendering details of the playhead have been moved into the Skin class and the playhead is now rendered within Qt instead of being drawn using a pixmap. Also, the margin of the PatternEditor was converted to a static constexpr. This way all classes and not just inherited ones can access it during initialization during which PatternEditorPanel is not fully initialized yet.

(Playhead might not be the proper term for the full-sized thingy. But there is already a cursor. So, I will just continue calling it playhead until something better shows up)