tweak selection highlighting
the border of selected patterns and notes do now use the same color (the one from the preferences). Also, the color of the selection lasso can now be altered as well.

As switching between active and inactive selection color was not working in the DrumPatternEditor as focus moved in and out of the widget, both the general structure of the background caching and the focusIn and focusOut policy was moved into the parent PatternEditor class.