PianoRollEditor: fixes, color theme + octaveRowColor
PianoRollEditor::createBackground() did not paint the actual pattern. This causes selection to not become inactive in case the focus leaves the widget

In addition, the alternate background color for a different octave was exposed in the preferences and the widget does now use the colors for the color theme. The design was a little bit simplified. The base octave hhad a slightly different color compared to the other octave highlights. At least I did only notice this in the code and since the base pitch default notes will be initialized with is highlighted, the base octave does now share the same color as all others.