Editors: align cursor hiding
Previously, the cursor was hidden when pressing a mouse key in SE but when clicking in DPR, NPR, and PRE. Hitting a key note bound to any move, selection, or value change operation did hide the cursor in SE and NPR but not in the others.

Now, the cursor is alway hidden on pressing a mouse key as well as hitting a key, like Shift.

Also, the setting the selected instrument in DPE is now done during mouse press instead of mouse click.