PatternEditorRuler: always show playhead
The overall UX of the PatternEditorRuler was altered. It does now always show the position of the playhead. But the playhead will only be full-height in case the currently selected pattern is amongst the playing patterns of the audio engine. In addition, the length of the active region is now solely determined by the longest pattern in the playing patterns list. It is thus decoupled from the pattern shown in the pattern editor and does always show the transport position.