AudioEngine: two consistent states
There have been some glitches when switching patterns during playback in pattern mode. Now UX is as follows. If user is in a short pattern a) and switches into a longer pattern b) while in selected patter mode, transport plays till the end of pattern b) and loops back to the beginning. In stacked pattern mode when playing pattern a) and adding b) to the next patterns (scheduling it to be played next), transport will be looped and playback of b) starts at the very beginning.

Also the transport position displayed by the cursor does no longer correspond to the actual transport position (AudioEngine::m_fTick) during playback but to the last position touched by AudioEngine::updateNoteQueue() (AudioEngine::m_fTick + lookahead). This has the drawback of introducing a noticable jump of the cursor when starting playback after a relocation (as the lookahead is reset during a relocation). On the other hand, it fixes the more severe issue that the cursor lags behind the enlisted notes and it was possible to stop the cursor right before a note with the note already being played back. This might not have been too obvious till now but it will when introducing the full-height cursor.

Internals: My previous goal was to keep the whole audio engine in one consistent state. But due to the lookahead this is not easily possible (without accounting for it at various places). Therefore there are now two consistent states: the transport information (most in TransportInfo apart from some helper variables) and are group of variables that - like m_nColumn, m_nPatternStartTick, m_nPatterSize, m_nPatterTickPosition - that are always in sync with AudioEngine::m_pPlayingPatterns.

I also fixed a bug occuring when resizing patterns in pattern mode while transport is rolling.