Song: move selected/stacked mode toggle into Song
The variable indicating whether selected or stacked pattern mode is used within Hydrogen was in a very strange state. It was located in the Preferences class and also written to disk as part as the hydrogen.conf file. On the other hand, it was also written into each .h2song file since the last 10 years. When loading a song the value was read and used to overwrite the preferences.

The variable is now an enum class to make the code more clear and easy to read and it is stored in Song::m_patternMode and accessed via the Hydrogen class much like action mode, loop mode, etc.

Also, the SongEditorPanel did not update properly when loading a different song. This has been fixed as well.