SongEditorPositionRuler: support active range
like the ruler in the pattern editor the one in the song editor is now aware of the range of supported transport positions too. The hover playhead is only shown within the bounds of the current song and the background of the position row beyond that is colored grey.

A new event type EVENT_SONG_SIZE_CHANGED was introduced