PatternEditor: show full playing patterns
In Song::PatternMode::Stacked playing notes not belonging to the currently selected pattern are drawn more faintly to indicate their presence and to ease tweaking patterns.

In this commit the display of the playing notes is extended to support the largest length among all playing patterns. If it exceeds the length of the currently selected pattern, the grid lines of and notes of the surplus columns will be drawn in window + windowText color.

Grid lines colors in PatternEditor are not static anymore in order to allow for them to be changed via the Preferences