PatternEditor: note property change
This commit moves the logic for editing and undoing note property changes via right mouse press and drag into the parent PatternEditor class. Now, it is possible in both the DrumPatternEditor and the PianoRollEditor to alter properties which makes their UX much more similar.

It also fixes a bug where undoing a property changed in the PRE was only possible if the pattern containing the note was selected