Drumkit export: store pan old-fashioned
Instead of using the most recent version of expressing the pan (from -1 to 1 with a single parameter; introduced in v1.2) the former version of having two variables "pan_L" and "pan_R" is used when storing drumkits.

This way drumkits created with version 1.2 of Hydrogen are still validate in older versions.