SongEditor: matching width + resize
the SongEditor and its ruler did not share the same length. As a consequence both went out of sync when scrolling to the end of the SE (with the default setting of max pattern sequences). Also, the maximum number of pattern sequences in the SongEditor required an update after changing in the PreferencesDialog. This will be now done immediately. Large tick numbers will now be fully displayed even for the lowest grid width possible