fix #1542: notes exceeding pattern length
Next to Pattern::__notes, which is a container for all notes within the pattern, and its accessor Pattern::get_notes() the new function Pattern::getAccessibleNotes() was introduced. Its purpose is to creating a fresh map similar to __notes but only containing accessible notes and provide it as shared pointer. The notes itself will not be copied but only their pointers. A note is accessible if its tick position within the pattern is smaller than the pattern length.

getAccessibleNotes() is not used at all places as a replacement for get_notes() since serialization, clearing all notes/instruments, and redo/undo should still work on all notes without any loss of information. Whether a note is accessible is primarly affecting notes playback, user interaction, visual rendering, and MIDI export.