Fix song modified handling
Previously a freshly opened song was marked modified right away. This has been fixed by ensuring the isModified flag is only set on user interactions. Missing flag placement has been added.

The state the isModified handling is in right now is not ideal though. It also marks the song modified when changing the state of a variable that is stored in the Preferences and not in the Song. But I think it would only lead to missing flags in future release and to a less intuitive UX if we would remove all flags for these widgets.