GMRockKit: fix holes
There were both tiny holes and overlaps between the layers of the GMRockKit. Since both of them were present and considering their small sizes I guess they were not introduced on purpose. I have aligned the layers the Sampler is less efficient in case a note is falling in such a hole.