PatternList: virtual pattern activation
When activating a pattern/adding it to a pattern list it is now checked whether the pattern was already present as part of a virtual one. If so, it will not be activated. In the same vain all individual activations will be removed if a virtual pattern is added that contains the already activated patterns.

Fixes #1527