Fail gracefully on missing playback track
As a user noted in #1536 Hydrogen does crash if the filename of the playback track points to non-existing file. The changes in here do, firstly, check for the existance of the file during song load and, secondly, perform a sanity check in the Sampler and deactivate the playback track in case a invalid file did slip through.

In addition, I did a little refactoring of the code in order to make the state of the playback track more clear from the perspective of the GUI (using one variable instead of juggling two).