AppVeyor [win]: lock qt5-tools-5.15.2-2
In order to circumvent #1559 the qt5-tools library is not used directly from the qt5 bundle of pacman but, instead, downgraded to version 5.15.2-2. This is only - at least hopefully - temporary measure till pacman offers a newer, working version of the package.

In addition, a bug in the artifact upload in the Windows image has been fixed.