Widgets: fix zero of WidgetWithInput
when altering the value of a `WidgetWithInput` - a `Rotary` or `Fader` - an increment is calculated based on the range of allowed values. During this process a tiny rounding error is introduced which itself does harm when setting values bigger or smaller than zero. But when trying to reset the widget to 0.0, it will be set to e.g. 3e-08 instead. To prevent this from happening, all values bigger than 1e-06 or smaller than -1e-06 are now truncated to 0.0