Make virtual keyboard independent of HearNewNotes
all notes inserted via the virtual keyboard are now played back. Previously this was only the case if the "Hear New Notes" button in the SongEditorPanel was checked (which was neither described in the tooltip nor in the docs). But the button did not properly deactivate the virtual keyboard either. It still could be used record notes. In case some people do miss this behaviour, I would suggest to add "disable/mute VK" option in the Preferences. But until then I would just keep it activated all the time as the corresponding keys do not serve another purpose within Hydrogen and should not annoy the user.

I removed two (now) redundant input arguments of Hydrogen::addRealtimeNote() and removed the section which activates playback of the added note only if it is located in the future. The latter is bypassed by MIDI input events and only used for the VK. But by design the VK shoud not be able to add notes which are not in the future. That was probably some legacy code.

As a consequence the VK can now also be used to play notes while transport is rolling in song mode without recording enabled.

fixes #1426