DrumPatternEditor: fix note deletion
in addOrDeleteNoteAction() the check whether or not to remove a note was done by (among others) comparing the pointer to a specific instrument with the one stored in the Note. However, this lead to problems when recording notes using the virtual keyboard while removing duplicates. Instead, the comparison is now based on the instrument ID.