MIDI: record NOTE OFF events
the code for recording NoteOn and NoteOff events in `MidiInput` was maintained separately and diverged over the years. As a result, the latter was picking a different pattern in case multiples are present and did not too the lookahead into account.

The NoteOff related parts are now integrated in Hydrogen::addRealtimeNote() (which cleans the related code by a margin).

Notes: the property `EventQueue::AddMidiNoteVector::b_noteExist` was dropped as both the core and the GUI were searching for duplicate notes. The former can be considered redundant as the latter is the instance responsible for actual removal of the note.
- A downside of the current implementation I only realize now: Recording incoming notes in a pattern is done by the GUI and not the core. That might become a problem when using Hydrogen as a plugin at some stage
- truncation of the new note length assigned when sending first a NOTEON and than a NOTEOFF event in case the length exceeds the pattern length was off. I did fix it but can test whether this has undesired side effects locally

fixes #1523