test: print cause of failing AE tests in console
the tests of the AudioEngine are located in the `AudioEngine` class in order to access private member variable without the need to create accessors for each of them. Whether or not a test succeeds is communicated by a return value. However, the console log in the CI only prints out the statement within the CPPUNIT ASSERT command which is the same for all AE tests. In order to determine the true reason for the failure one has to run the tests locally with the -V argument.

To circumvent this, have more verbose results in the CI, and to catch a transient failure in the note enqueuing that seems to popup in the CI but which I can not reproduce locally, the AE test do print their errors directly using `qDebug()` instead of Hydrogens custom log queue. This way the cause of the problem should also appear in the console log of the CI.