XSD: drop default values
there were a number of default values for various parameters present in the XSD schema files. In principle, this is a nice idea. However, a number of them were not properly updated and are out-of-sync with the real default values assigned in the code. Also, there is no way - at least to my knowledge - to harness the default values of the XSD files and to use them as default values in the C++ code. To not cause confusion, I all default values were dropped.

Implements #1197