audio driver selection
resort to NullDriver in case the driver picked by the user could not be started instead of falling back to a different one. For the latter behavior the "Auto" driver can be used.

A problem that occurred with the previous behavior was that a different driver was started by Hydrogen unbeknownst to the user. Now, the Preferences dialog will prompt an error popup when not able to restart the selected driver and prompt a warning dialog in case the OK button was clicked while NullDriver is present in the AudioEngine. By letting the audio driver selection fail temporarily the whole selection process becomes more clear.