Note: fix note construction from tmp file (#1587)
when duplicating a pattern, the duplicated one is written to a temp file and discarded. The pattern actually showing up in the GUI was constructed using the temp file. During this construction the `__instrument` member of the associated notes is assigned in `Note::map_instrument()` instead of being provided to the constructor itself. However, the function forgets to properly initialized both the `__adsr` and `__layers_selected` of the note causing it to fail playback in the current session.