Update license setting in GUI
both the property dialogs of the song and the drumkit does now fully support setting the new License class.

Combo boxes will be provided to choose between one of the definitions in License::LicenseType enum. If License::Other is selected, a QLineEdit will be shown using which the user can set a custom license