SoundLibraryPropertiesDialog: content table
A new table listing all samples associated to the drumkit including their instrument and component name as well as their license was introduced. In case of a sample differs from the one of the overall drumkit the row is highlighted in red.

In addition, the box containing the drumkit image will now be hidden if no drumkit image was specified and an error is logged in case the specified image could not be loaded.