rework License class
Instead of making the License::m_license License::Licensetype the primary vehicle to deal with licensing information within Hydrogen it now has just a supportive role. Problem was that the user could assign arbitrary attribution text that would have been lost on the predefined licenses.

In all property dialogs both the combo box to select from a predefined license and a line edit showing the string that is stored on disk are both shown.