License: add GPL license notice
in order to comply with the GPL license, a copy of the license notice has to be included. This was never the case in any of our drumkits up to now. Regardless of the fact that most of them are licensed as GPL.

Now, whenever a drumkit or a song is licensed as GPL a license notice is included as comment in the drumkit.xml or .h2song file