SoundLibraryProperties: handle song instrument list
Previously the property dialog invoked by right-clicking the currently loaded drumkit in the SoundLibrary and by clicking menu > Drumkit > Properties were identially. However, this lead to confusion as the instrument list shown in the content table was always loaded from the drumkit. Now, the instrument list found in the current song is shown in case the dialog was opened via the main menu.

Importing an instrument of a different kit by click-dragging adds the corresponding samples of the content table in the dialog.

Some functions from DrumPatternEditor working solely on the drumkit components of a song were moved into the song class.